ASCII

+++ lab2.c
@@ -175,7 +175,7 @@ int rws = 1;
 
 ///////////////////////////////////////////////////////////////////
 
-char usbkey_to_ascii(uint8_t keycode, uint8_t modifiers)
+char usbkey_to_ascii(int keycode, int modifiers)
 {
     // Letters: keycodes 0x04 to 0x1d represent 'a' to 'z'
     if (keycode >= 0x04 && keycode <= 0x1d) {
